added codes to add sites

+++ app/Controllers/logout.php
@@ -2,6 +2,12 @@
 
 use App\Core\Session;
 
-Session::destroy();
-header("location: /");
-exit();
+if (isset($_POST['submit']) && $_SERVER['REQUEST_METHOD'] === 'POST') {
+    Session::destroy();
+    header("location: /");
+    exit();
+} else {
+    echo "Page not found.<br>";
+    echo "<a href='/'>Go Home</a>";
+
+}
